Modify Lib.getTargetArray
